feat(app):  Implemented shimmer on DishDetailsScreen, fixed errors, added image loading to DishDetailsScreen

+++ Recipely/Recipely/Services/NetworkService/DTO/NutrientInfoDTO.swift
@@ -6,5 +6,5 @@ import Foundation
 /// Структура, представляющая питательное вещество.
 struct NutrientInfoDTO: Codable {
     /// Количество питательного вещества.
-    let quantity: Double
+    let quantity: Float
 }
